@@ -1,57 +1,83 @@
 {
 	"name": "vue3-slide-up-down",
-	"version": "1.2.5",
-	"description": "Vue 3 version of the vue-slide-up-down package",
-	"source": "src/index.js",
-	"amdName": "Vue3SlideUpDown",
-	"main": "dist/vue3-slide-up-down.cjs.js",
-	"umd:main": "dist/vue3-slide-up-down.umd.js",
-	"module": "dist/vue3-slide-up-down.es.js",
-	"files": [
-		"dist",
-		"*.ts"
-	],
+	"version": "2.0.0",
+	"author": "southcoastweb",
+	"homepage": "https://southcoastweb.co.uk",
 	"scripts": {
-		"build": "vite build -c vite-build.config.js",
-		"dev": "vite build -c vite-build.config.js --watch",
-		"demo": "vite",
-		"demo:build": "vite build",
-		"demo:deploy": "gh-pages -d demo/dist"
+		"dev": "vite",
+		"build": "vite build && tsc --emitDeclarationOnly",
+		"test:unit": "vitest --dom",
+		"lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+		"format": "prettier --write src/",
+		"prepublishOnly": "npm run lint && npm run format && npm run build"
 	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/craigrileyuk/vue3-slide-up-down.git"
+	"type": "module",
+	"main": "./dist/vue3-slide-up-down.cjs",
+	"module": "./dist/vue3-slide-up-down.js",
+	"types": "./dist/types/index.d.ts",
+	"exports": {
+		".": {
+			"require": "./dist/vue3-slide-up-down.cjs",
+			"import": "./dist/vue3-slide-up-down.js",
+			"browser": "./dist/vue3-slide-up-down.umd.cjs",
+			"types": "./dist/types/index.d.ts"
+		}
 	},
-	"keywords": [
-		"vue",
-		"vue3",
-		"slideup",
-		"slidedown",
-		"reflow"
+	"files": [
+		"dist"
 	],
-	"author": "Craig Riley",
-	"license": "MIT",
-	"bugs": {
-		"url": "https://github.com/craigrileyuk/vue3-slide-up-down/issues"
-	},
-	"homepage": "https://github.com/craigrileyuk/vue3-slide-up-down#readme",
-	"devDependencies": {
-		"@vitejs/plugin-vue": "^1.2.0",
-		"@vue/compiler-sfc": "^3.1.0",
-		"autoprefixer": "^10.3.1",
-		"eslint": "^7.32.0",
-		"eslint-plugin-jsdoc": "^36.0.7",
-		"eslint-plugin-vue": "^7.16.0",
-		"gh-pages": "^3.1.0",
-		"postcss": "^8.3.6",
-		"prettier": "^2.4.1",
-		"prismjs": "^1.23.0",
-		"sass": "^1.30.0",
-		"tailwindcss": "^2.2.7",
-		"vite": "^2.4.4"
-	},
 	"peerDependencies": {
 		"vue": "^3.0.0"
 	},
-	"typings": "./index.d.ts"
+	"devDependencies": {
+		"@rushstack/eslint-patch": "^1.2.0",
+		"@testing-library/vue": "^7.0.0",
+		"@types/jsdom": "^21.1.0",
+		"@types/node": "^18.14.0",
+		"@vitejs/plugin-vue": "^4.0.0",
+		"@vue/eslint-config-prettier": "^7.1.0",
+		"@vue/eslint-config-typescript": "^11.0.2",
+		"@vue/test-utils": "^2.3.0",
+		"@vue/tsconfig": "^0.1.3",
+		"autoprefixer": "^10.4.14",
+		"eslint": "^8.34.0",
+		"eslint-plugin-vue": "^9.9.0",
+		"happy-dom": "^9.8.4",
+		"jsdom": "^21.1.0",
+		"npm-run-all": "^4.1.5",
+		"postcss": "^8.4.22",
+		"prettier": "^2.8.4",
+		"tailwindcss": "^3.3.1",
+		"typescript": "^5.0.2",
+		"vite": "^4.1.3",
+		"vite-plugin-dts": "^2.0.2",
+		"vitest": "^0.30.1",
+		"vue-tsc": "^1.1.5"
+	},
+	"browserslist": [
+		"> 1%",
+		"last 2 versions",
+		"not dead"
+	],
+	"bugs": {
+		"url": "https://github.com/southcoastweb/vue3-slide-up-down/issues"
+	},
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/southcoastweb/vue3-slide-up-down.git"
+	},
+	"keywords": [
+		"Vue",
+		"Vue3",
+		"show",
+		"hide",
+		"slide",
+		"collapse",
+		"transition",
+		"slideup",
+		"slidedown",
+		"up",
+		"down"
+	]
 }
